+++ java/clients/src/test/java/sleeper/clients/api/SleeperClientIngestIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.clients.cli;
+package sleeper.clients.api;
 
 import com.amazonaws.services.sqs.model.Message;
 import org.junit.jupiter.api.BeforeEach;
