+++ java/clients/src/test/java/sleeper/clients/api/SleeperClientTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.clients.cli;
+package sleeper.clients.api;
 
 import com.google.common.collect.ImmutableMap;
 import org.junit.jupiter.api.Test;
